requirements.txt
./wheels/yaml_reader_lib-0.1.0-py3-none-any.whl
yaml_reader_lib==0.1.0
